+++ CHANGELOG.md
@@ -7,11 +7,13 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Gamepad support for the map editor's absolute/relative terrain height brushes and natural resource spawning/removal brushes
+- Optional support for the FPP mod
 
 ### Fixed
 
 - Gamepad UI cursor now correctly defaults to the select tool
 - Natural resource spawning/removal brushes no longer immediately apply the instant they were selected
+- Mouse/keyboard can now be used for building placement and all area tools alongside the controller
 
 ## [1.1.2.0.3] - 2026-08-27